requirements.txt
python-dotenv
spond